weather_agent/requirements.txt
google-adk[a2a]>=1.0.0
python-dotenv>=1.0.1
